split single cli to cli and lib for using in other projects
