+++ .gitignore
@@ -1,1 +1,1 @@
-bin
+dist/
