+++ tsconfig.json
@@ -4,7 +4,7 @@
     "lib": ["esnext"],
     "module": "commonjs",
     "declaration": true,
-    "outDir": "./bin",
+    "outDir": "./dist",
     "strict": true
   },
   "include": ["src"],
